batt_man.sh: Add support for ADP5061
This patch adds support for the new ADP5061 driver.

If the driver is installed, then the charger status is read by using the
power_supply status property routine. The values returned by this attibute
are Charging, Not charging, Full, Discharging and Unknown.

If the driver is not installed, then the status is retrieved via the i2cget
routine. The values returned were grouped to be consistent with the
power_supply status attribute.

Signed-off-by: Stefan Popa <stefan.popa@analog.com>