EnumElement: added cmd_enum, cmd_enum_timer, cmd_enum_sep
Using cmd_enum property (instead of enum), this element can dynamically
change it's elements based on a command. cmd_enum_timer will set the
update frequency by connecting it to one of the menu timers, and
cmd_enum_sep will change the way the output is split from the output of
the command in cmd_enum.

Element example:
      {
              "description": "Files ",
              "type": "edit",
              "subtype": "enum",
              "cmd_read": "cat /tmp/file.txt",
              "cmd_read_timer": "read_once",
              "cmd_write": "echo $EDIT_VAL > /tmp/file.txt",
              "cmd_enum": "ls /usr/local/bin",
	      "cmd_enum_timer": "read_once",
              "cmd_enum_sep": " "
        }

Signed-off-by: Adrian Suciu <adrian.suciu@analog.com>